+++ src/main/java/svenhjol/charmony/runestones/common/features/runestones/Helpers.java
@@ -3,37 +3,23 @@ package svenhjol.charmony.runestones.common.features.runestones;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.RegistryAccess;
 import net.minecraft.core.registries.Registries;
-import net.minecraft.network.protocol.game.ClientboundChangeDifficultyPacket;
-import net.minecraft.network.protocol.game.ClientboundPlayerAbilitiesPacket;
-import net.minecraft.network.protocol.game.ClientboundRespawnPacket;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.server.level.ServerLevel;
-import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.tags.TagKey;
-import net.minecraft.util.Mth;
 import net.minecraft.util.RandomSource;
-import net.minecraft.world.entity.Entity;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelAccessor;
 import net.minecraft.world.level.biome.Biome;
-import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.levelgen.structure.Structure;
-import net.minecraft.world.phys.Vec3;
 import svenhjol.charmony.api.RunestoneLocation;
-import svenhjol.charmony.core.base.Log;
 import svenhjol.charmony.core.helpers.TagHelper;
 import svenhjol.charmony.runestones.RunestonesMod;
 
-import javax.annotation.Nullable;
 import java.util.List;
-import java.util.Locale;
 import java.util.Optional;
 
 public final class Helpers {
-    private static final Log LOGGER = new Log(RunestonesMod.ID, "Helpers");
-
     public static final ResourceLocation SPAWN_POINT_ID = RunestonesMod.id("spawn_point");
     public static final ResourceLocation EMPTY_ID = RunestonesMod.id("empty");
     public static final RunestoneLocation SPAWN_POINT = new RunestoneLocation(RunestoneLocation.Type.Player, SPAWN_POINT_ID);
@@ -118,66 +104,6 @@ public final class Helpers {
         var registry = registryAccess.lookupOrThrow(tag.registry());
         return TagHelper.getValues(registry, tag)
             .stream().map(registry::getKey).toList();
-    }
-
-    /**
-     * @see ServerPlayer#teleport
-     */
-    public static void changeDimension(ServerPlayer serverPlayer, ServerLevel newDimension, Vec3 pos) {
-        serverPlayer.isChangingDimension = true;
-        var connection = serverPlayer.connection;
-        var currentDimension = serverPlayer.serverLevel();
-        var levelData = newDimension.getLevelData();
-
-        connection.send(new ClientboundRespawnPacket(serverPlayer.createCommonSpawnInfo(newDimension), (byte)3));
-        connection.send(new ClientboundChangeDifficultyPacket(levelData.getDifficulty(), levelData.isDifficultyLocked()));
-
-        var playerList = currentDimension.getServer().getPlayerList();
-        playerList.sendPlayerPermissionLevel(serverPlayer);
-        currentDimension.removePlayerImmediately(serverPlayer, Entity.RemovalReason.CHANGED_DIMENSION);
-        serverPlayer.unsetRemoved();
-
-        var yRot = serverPlayer.getYRot();
-        var xRot = serverPlayer.getXRot();
-
-        serverPlayer.setServerLevel(newDimension);
-        connection.teleport(pos.x(), pos.y(), pos.z(), yRot, xRot);
-        connection.resetPosition();
-
-        newDimension.addDuringTeleport(serverPlayer);
-
-        connection.send(new ClientboundPlayerAbilitiesPacket(serverPlayer.getAbilities()));
-        connection.send(new ClientboundPlayerAbilitiesPacket(serverPlayer.getAbilities()));
-
-        playerList.sendLevelInfo(serverPlayer, newDimension);
-        playerList.sendAllPlayerInfo(serverPlayer);
-        playerList.sendActivePlayerEffects(serverPlayer);
-        
-        serverPlayer.lastSentExp = -1;
-        serverPlayer.lastSentHealth = -1.0f;
-        serverPlayer.lastSentFood = -1;
-    }
-
-    @Nullable
-    public static BlockPos getSurfacePos(Level level, BlockPos pos, int startAtHeight) {
-        int surface = 0;
-
-        for (int y = startAtHeight; y >= 0; --y) {
-            BlockPos n = new BlockPos(pos.getX(), y, pos.getZ());
-            if (level.isEmptyBlock(n)
-                && !level.isEmptyBlock(n.below())
-                && !level.getBlockState(n.below()).is(Blocks.LAVA)) {
-                surface = y;
-                break;
-            }
-        }
-
-        if (surface == 0) {
-            LOGGER.warn("Failed to find a surface value to spawn the player");
-            return null;
-        }
-
-        return new BlockPos(pos.getX(), surface, pos.getZ());
     }
 
     public static String localeKey(RunestoneLocation location) {