+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 1.7.0
+
+- Fix runestones facing the same way when placed by the game.
+- Add new seed calculation based on runestones position to avoid duplication on the Y axis.
+- Add configuration for stone circle debris and to add runestone excavation to debris.
+- Add many more sacrifice items for each type of runestone.
+- Rebase teleportation on the the Charmony teleporter class.
+- Move runestones and stone circle definitions into their own init classes.
+- Move API consumer code into the register classes.
+- Move advancement language strings into the correct asset folder.
+
 ## 1.6.1
 
 - Update to use rune dictionary library.